Add files via upload
Objective: 
Conduct a comparison analysis of iPhone 15 sales and iPhone 14 sales. 
 
Datasets: 
1.	fact_sales_Iphone14.csv 
2.	fact_sales_Iphone15.csv 
3.	dim_stores.csv 
 
Tasks: 
Here is a brief overview of the workflow for the task: 
1.	Review the data (fact_sales_Iphone14 and fact_sales_Iphone15). 
2.	Metrics to compare - iPhone 14 sales and iPhone 15 sales. 
3.	Calculate the variance (absolute difference value) between the iPhone 15 sales and iPhone 14 sales for the identified metrics. ABS([Iphone 15 Sales]-[Iphone 14 Sales]). 
4.	Calculate the variance (in percentage) between the iPhone 15 sales and iPhone 14 sales for the identified metrics. [([Iphone 15 Sales]-[Iphone 14 Sales])/[Iphone 14 Sales]. 
5.	Provide a complete report of the difference in numbers between the iPhone 15 sales and iPhone 14 sales data across the top 10 countries. 
 
Please ensure that you thoroughly analyze the datasets and provide a comprehensive report of the differences. 
Attaching a screenshot from a sample report. It will give you a better idea of what it looks like and how the final output should be presented.